refactor: merge tests for authorization parameter validator
