--- src/Application/Authorization/ParameterValidation/IAuthorizationParameterValidator.cs
@@ -1,14 +1,0 @@
-﻿using Core.Client;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace Application.Authorization.ParameterValidation
-{
-    public interface IAuthorizationParameterValidator
-    {
-        Task<AuthorizationParameterValidationResult> ValidateClientId(string clientId);
-        AuthorizationParameterValidationResult ValidateRedirectUrl(string redirectUrl, IEnumerable<RedirectUrl> registeredRedirectUrls);
-        AuthorizationParameterValidationResult ValidateResponseType(string responseType, ClientType clientType);
-        Task<AuthorizationParameterValidationResult> ValidateScope(string scope);
-    }
-}
