--- src/Application/Authorization/ParameterValidation/ResponseType.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Application.Authorization.ParameterValidation
-{
-    public class ResponseType
-    {
-        public const string Code = "code";
-        public const string IdToken = "id_token";
-        public const string AccessToken = "token";
-        public const string None = "none";
-    }
-}
