+++ src/Core/Authorization/ParameterValidation/ResponseType.cs
@@ -1,0 +1,9 @@
+﻿namespace Core.Authorization.ParameterValidation;
+
+public class ResponseType
+    {
+    public const string Code = "code";
+    public const string IdToken = "id_token";
+    public const string AccessToken = "token";
+    public const string None = "none";
+}
